+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - [9985](https://github.com/vegaprotocol/vega/issues/9985) - Add update margin mode transaction. 
 - [9982](https://github.com/vegaprotocol/vega/issues/9982) - Remove fees and minimal transfer amount from vested account
 - [9955](https://github.com/vegaprotocol/vega/issues/9955) - Add data node subscription for transaction results.
+- [9825](https://github.com/vegaprotocol/vega/issues/9825) - Remove quadratic slippage.
 
 ### 🐛 Fixes